+++ src/main/java/com/zematix/jworldcup/backend/configuration/CachingConfig.java
@@ -23,7 +23,7 @@ public class CachingConfig {
 	public static final String CACHE_FAVOURITE_KNOCKOUT_TEAMS = "favouriteKnockoutTeams";
 
     @Bean
-    public CacheManager cacheManager() {
+    CacheManager cacheManager() {
         CaffeineCacheManager cacheManager = new CaffeineCacheManager(
         		CACHE_USER_BY_LOGIN_NAME, CACHE_USER_OF_EVENT, CACHE_USER_GROUPS, 
         		CACHE_EVENT_START_TIME, CACHE_EVENT_KNOCKOUT_START_TIME, CACHE_EVENT_END_TIME,