+++ src/main/java/com/zematix/jworldcup/backend/configuration/JacksonConfig.java
@@ -20,7 +20,7 @@ public class JacksonConfig {
    */
 	@Bean
 	@Primary
-	public ObjectMapper objectMapper(Jackson2ObjectMapperBuilder builder) {
+	ObjectMapper objectMapper(Jackson2ObjectMapperBuilder builder) {
 		ObjectMapper objectMapper = builder.createXmlMapper(false)
 				.serializers(new LocalDateTimeUTCSerializer(LocalDateTime.class))
 				.build();