+++ src/main/java/com/zematix/jworldcup/backend/dao/RoundDao.java
@@ -10,13 +10,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.querydsl.jpa.impl.JPAQuery;
 import com.zematix.jworldcup.backend.entity.Event;
-import com.zematix.jworldcup.backend.entity.QRole;
 import com.zematix.jworldcup.backend.entity.QRound;
-import com.zematix.jworldcup.backend.entity.Role;
 import com.zematix.jworldcup.backend.entity.Round;
 
 /**
- * Database operations around {@link Role} entities.
+ * Database operations around {@link Round} entities.
  */
 @Component
 @Transactional
@@ -28,8 +26,7 @@ public class RoundDao extends DaoBase {
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public List<Round> getAllRounds() {
 		TypedQuery<Round> query = getEntityManager().createNamedQuery("Round.findAll", Round.class);
-		List<Round> roles = query.getResultList();
-		return roles;
+		return query.getResultList();
 	}
 
 	/**
@@ -46,32 +43,10 @@ public class RoundDao extends DaoBase {
 		
 		QRound qRound = QRound.round;
 		JPAQuery<Round> query = new JPAQuery<>(getEntityManager());
-		rounds = query.from(qRound)
+		return query.from(qRound)
 			.where(qRound.event.eventId.eq(eventId))
 			.orderBy(qRound.roundId.asc())
 			.fetch();
-		
-		return rounds;
-	}
-	
-	/**
-	 * Return found {@link Role} instance which matches the given role string value. 
-	 * Otherwise {@code null} is returned.
-	 * 
-	 * @param - sRole - searched role string
-	 * @return found {@link Role} entity instance or {@code null} if not found 
-	 */
-	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-	public Role findRoleByRole(String sRole) {
-		Role role = null;
-		
-		QRole qRole = QRole.role1;
-		JPAQuery<Role> query = new JPAQuery<>(getEntityManager());
-		role = query.from(qRole)
-			.where(qRole.role.eq(sRole))
-		  .fetchOne();
-		
-		return role;
 	}
 	
 }