+++ src/main/java/com/zematix/jworldcup/backend/emun/DictionaryEnum.java
@@ -1,0 +1,12 @@
+package com.zematix.jworldcup.backend.emun;
+
+import com.zematix.jworldcup.backend.entity.Dictionary;
+
+/**
+ * Contains possible types of {@link Dictionary}
+ */
+public enum DictionaryEnum {
+	ROLE,
+	USER_NOTIFICATION,
+	USER_STATUS
+}