+++ src/main/java/com/zematix/jworldcup/backend/emun/UserStatusEnum.java
@@ -1,0 +1,10 @@
+package com.zematix.jworldcup.backend.emun;
+
+/**
+ * Contains possible types of UserStatus
+ */
+public enum UserStatusEnum {
+	NORMAL,
+	CANDIDATE,
+	LOCKED
+}